refactor(providers): split notifications providers out of app_providers.dart
Fourth batch of #4506. Moves the entire 10-entity notifications cluster
to `lib/providers/notifications_providers.dart`:

- firebaseMessagingProvider, firebaseOnMessageProvider
- notificationServiceProvider
- notificationPreferencesStoreProvider (Hive-backed)
- pushNotificationServiceProvider (auth/session-gated)
- _updateNotificationPreferencesSafely (file-private helper)
- notificationPreferencesDirtySyncBridgeProvider (retry state machine)
- notificationPreferencesServiceProvider
- pushNotificationSync (@Riverpod keepAlive — eager FCM-and-coordinator)
- notificationServiceEnhanced (@riverpod — see notes)

The two preferences-side providers reference each other via lazy
`ref.read` callbacks; they must stay co-located. The file-private
`_updateNotificationPreferencesSafely` helper moves with them.

`app_providers.dart` adds `export 'notifications_providers.dart';`. No
back-import needed — no remaining provider in app_providers.dart
consumes a notifications symbol (the collaborator block that was
interleaved at lines 195-235 stays in app_providers.dart untouched).

The new file imports `app_providers.dart` for `authServiceProvider`,
`relayStatisticsServiceProvider`, `profileRepositoryProvider`,
`videoEventServiceProvider`, and imports `database_provider.dart`,
`environment_provider.dart`, `nostr_client_provider.dart` directly
for their respective providers (already-separated files).

Removed now-unused imports from app_providers.dart: firebase_messaging,
flutter/foundation (kIsWeb), notification_preferences (model),
notification_preferences_service, notification_service,
push_notification_service, push_notification_session_coordinator. Kept
notification_service_enhanced.dart because app_providers.dart still
calls `NotificationServiceEnhanced.instance.clearAllData()` from the
sign-out cleanup callback.

Codegen moved: pushNotificationSyncProvider and
notificationServiceEnhancedProvider relocated from app_providers.g.dart
into notifications_providers.g.dart.

Behavior preserved verbatim. Two follow-up issues to file separately:
1. Extract NotificationPreferencesDirtySyncCoordinator service — the
   95-line state-machine closure inside the bridge provider is
   repository-layer logic embedded in the providers layer.
2. Resolve dead notificationServiceEnhancedProvider — it has zero lib
   consumers; the singleton init never fires in production.

Refs #4506, #4339 Wave 2
